pyproject.toml
[tool.poetry]
name = "bot_project"
version = "0.1.0"
description = ""
authors = ["Sergei <yasen20008@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.7"
ephem = "^4.0.0"
python-telegram-bot = "^13.7"
psycopg2-binary = "^2.9.1"
SQLAlchemy = "^1.4.25"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
